@@ -1,30 +1,44 @@
 package capstone.android;
-import customstorage.GridAdapter;
-import customstorage.Storage;
+import customized.customized.background.BackgroundProcess2;
+import customized.customized.listviewgroup1.ListViewAdapter;
+import customized.customized.data.Storage;
 import gun0912.tedbottompicker.TedBottomPicker;
 
+
+import android.content.Context;
 import android.content.Intent;
 
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.Typeface;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.Button;
-import android.widget.GridView;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
+import android.widget.ListView;
+import android.widget.TextView;
 import android.widget.Toast;
 
+import java.io.FileOutputStream;
+import java.io.ObjectOutputStream;
 import java.util.ArrayList;
+import java.util.List;
 
 
-public class MainActivity extends AppCompatActivity {
-
-    private BackPressCloseHandler backPressCloseHandler;
-    private static final int[] checkcode = {111,222};
+public class MainActivity extends AppCompatActivity{
+    private final int checkcode[] ={111,222,333};
+    private final String TAG = "MainActivity";
+    private BackPressCloseHandler backPressCloseHandler = new BackPressCloseHandler(this);
     private Storage storage;
-    private ArrayList<Button> buttonList = new ArrayList<Button>();
-    private GridView gridView;
-    private GridAdapter adapter;
+    private Button[] buttonList;
+    private TextView[] textviewList;
+    private ListView listView;
+    private ListViewAdapter adapter = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -35,164 +49,178 @@ public class MainActivity extends AppCompatActivity {
         gallery();
         Cam();
         Change();
+        this.listView.setOnItemClickListener(onItemClickListener);
     }
 
     private void setting() {
         this.storage = new Storage();
-        this.buttonList.add((Button)findViewById(R.id.btn_gallery));
-        this.buttonList.add((Button)findViewById(R.id.btn_cam));
-        this.buttonList.add((Button)findViewById(R.id.btn_change));
-        this.gridView = (GridView) findViewById(R.id.gridView1);
-        this.adapter = null;
-        this.backPressCloseHandler = new BackPressCloseHandler(this);
-        Log.d("MainActivity0",
-                "storage :"+this.storage.toString()+"\n"+
-                "buttonlist(1) :"+this.buttonList.get(0).toString()+"\n"+
-                "buttonlist(2) :"+this.buttonList.get(1).toString()+"\n"+
-                "buttonlist(3) :"+this.buttonList.get(2).toString()+"\n"+
-                "gridView :"+this.gridView.toString()+"\n"+
-                "backPressCloseHandler :"+this.backPressCloseHandler.toString()+"\n"
-
-        );
+        this.buttonList = new Button[]{(Button)findViewById(R.id.btn_gallery),(Button)findViewById(R.id.btn_cam),(Button)findViewById(R.id.btn_change)};
+        this.textviewList = new TextView[]{(TextView)findViewById(R.id.tV),(TextView)findViewById(R.id.textCam),(TextView)findViewById(R.id.textGal)};
+        this.listView = (ListView) findViewById(R.id.list_view);
+        Typeface typeFace_eng = Typeface.createFromAsset(getAssets(), "fonts/cour.ttf");
+        Typeface typeFace_kor = Typeface.createFromAsset(getAssets(), "fonts/NanumGothic.ttf");
+        this.buttonList[0].setTypeface(typeFace_eng);
+        this.buttonList[1].setTypeface(typeFace_eng);
+        this.buttonList[2].setTypeface(typeFace_kor);
+        this.textviewList[0].setTypeface(typeFace_eng);
+        this.textviewList[1].setTypeface(typeFace_eng);
+        this.textviewList[2].setTypeface(typeFace_eng);
     }
+
     private void gallery() {
-        buttonList.get(0).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                /*
-                Intent intent = new Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
-                intent.setType("image/*");
-                intent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE,true);
-                intent.setAction(Intent.ACTION_GET_CONTENT);
-                startActivityForResult(Intent.createChooser(intent, "Select Picture"),  checkcode[1]);*/
-                TedBottomPicker bottomSheetDialogFragment  = new TedBottomPicker.Builder(MainActivity.this)
-                        .setOnMultiImageSelectedListener(new TedBottomPicker.OnMultiImageSelectedListener(){
-                            @Override
-                            public void onImagesSelected(ArrayList<Uri> uriList){
-                                ArrayList<String> atemp = new ArrayList<String>();
-                                for(int i=0;i<uriList.size();i++){
-                                    Log.d("TedBottomPicker","temp0 :" +uriList.get(i).getPath().toString());
-                                    MainActivity.this.storage.setApath(i,uriList.get(i).getPath().toString());
-                                }
-                                MainActivity.this.adapter = new GridAdapter(MainActivity.this, MainActivity.this.storage.getApathswithAlist(),10);
-                                MainActivity.this.gridView.setAdapter(MainActivity.this.adapter);
-                                /*
-                                stemp = uriList.get(0).getPath().toString();
-                                storage.setabsolutePath(0,stemp);
-                                Bitmap bitmap = BitmapFactory.decodeFile(storage.getabsolutePath(0));
-                                int nh = (int) (bitmap.getHeight() * (1024.0 / bitmap.getWidth()));
-                                Bitmap scaled = Bitmap.createScaledBitmap(bitmap, 1024, nh, true);
-                                imgView.setImageBitmap(scaled);
-                                Toast.makeText(MainActivity.this,stemp+"이미지 선택 완료", Toast.LENGTH_SHORT).show();
-                                */
+        buttonList[0].setOnClickListener(v -> {
+            Intent intent = new Intent(MainActivity.this,GalleryActivity.class);
+            intent.putExtra("currentSize",0);
+            startActivityForResult(intent,222);
+            /*
+            TedBottomPicker bottomSheetDialogFragment  = new TedBottomPicker.Builder(MainActivity.this)
+                    .setPreviewMaxCount(100)
+                    .setSelectMaxCount(20-storage.getSize())
+                    .showTitle(false)
+                    .showCameraTile(false)
+                    .setEmptySelectionText("선택하지 않았습니다.")
+                    .setSelectMaxCountErrorText("최대 20개 까지입니다")
+                    .setCompleteButtonText("선택 완료")
+                    .setEmptySelectionText("No Select")
+                    .setOnMultiImageSelectedListener(new TedBottomPicker.OnMultiImageSelectedListener(){
+                        @Override
+                        public void onImagesSelected(ArrayList<Uri> uriList){
+                            ArrayList<String> atemp = new ArrayList<String>();
+                            for(int i=0;i<uriList.size();i++){
+                                Log.d("TedBottomPicker","temp0 :" +uriList.get(i).getPath());
+                                MainActivity.this.storage.addImageData(uriList.get(i).getPath());
                             }
-                        })
-                        .setSelectMaxCount(20)
-                        .showTitle(false)
-                        .setCompleteButtonText("선택")
-                        .setEmptySelectionText("No Select")
-                        .create();
-                bottomSheetDialogFragment.show(getSupportFragmentManager());
-            }
+                            MainActivity.this.adapter = new ListViewAdapter(MainActivity.this, MainActivity.this.storage,10);
+                            MainActivity.this.listView.setAdapter(MainActivity.this.adapter);
+                            visible();
+                        }
+                    })
+                    .create();
+            bottomSheetDialogFragment.show(getSupportFragmentManager());
+            */
         });
     }
+
     private void Cam(){
-        buttonList.get(1).setOnClickListener(new View.OnClickListener(){
-            @Override
-            public void onClick(View v){
-                Intent intent = new Intent(MainActivity.this, CaptureActivity.class);
-                startActivityForResult(intent,checkcode[0]);
-            }
+        buttonList[1].setOnClickListener(v -> {
+            Intent intent = new Intent(MainActivity.this, CaptureActivity.class);
+            startActivityForResult(intent,checkcode[0]);
         });
     }
 
     private void Change(){
-        buttonList.get(2).setOnClickListener(new View.OnClickListener(){
-            @Override
-            public void onClick(View v){
-                Intent intent = new Intent(MainActivity.this, OcrActivity.class);
-                intent.putExtra("customstorage", storage);
-                startActivity(intent);
+        buttonList[2].setOnClickListener(v -> {
+            Intent intent = new Intent(MainActivity.this, OcrActivity.class);
+            try {
+                writeObjectFile();
+            } catch (Exception e){
+                e.printStackTrace();
             }
+            startActivity(intent);
         });
+        //buttonList[2].setVisibility(View.GONE);
     }
 
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
+        ArrayList<String> absolList;
+        int id;
+        String cropedPath;
         if (requestCode == checkcode[0]) {
                 switch(resultCode){
                     case RESULT_OK:
-                        String temp = data.getStringExtra("camfilepath");
-                        /*
-                        storage.setabsolutePath(0,temp);
-                        Bitmap bitmap = BitmapFactory.decodeFile(temp);
-                        int nh = (int) (bitmap.getHeight() * (1024.0 / bitmap.getWidth()));
-                        Bitmap scaled = Bitmap.createScaledBitmap(bitmap, 1024, nh, true);
-                        this.imgView.setImageBitmap(scaled);
-                        */
-                        ArrayList<String> atemp = new ArrayList<String>();
-                        atemp.add(temp);
-                        this.adapter = new GridAdapter(MainActivity.this,atemp,10);
-                        break;
-                    default:
-                        Toast.makeText(this, "취소되었습니다.", Toast.LENGTH_SHORT).show();
-                        break;
-                }
+                        absolList = data.getStringArrayListExtra("CabsoulteFilePaths");
+                        for(int i = 0; i<absolList.size();i++){
+                            this.storage.addImageData(absolList.get(i));
+                        }
+                        try{
+                            this.adapter = new ListViewAdapter(MainActivity.this,this.storage,10);
+                            this.listView.setAdapter(this.adapter);
+                            visible();
 
-            } /*else if(requestCode == checkcode[1]){
-                switch(resultCode){
-                    case RESULT_OK:
-                        Uri uri = data.getData();
-                        Log.d("MainActivity1",uri.toString());
-                        storage.setabsolutePath(0,getImageApath(uri));
-                        Log.d("MainActivity2","이미지 절대 경로 : "+ storage.getabsolutePath(0));
-                        try {
-                            Bitmap bitmap = MediaStore.Images.Media.getBitmap(getContentResolver(), uri);
-                            int nh = (int) (bitmap.getHeight() * (1024.0 / bitmap.getWidth()));
-                            Bitmap scaled = Bitmap.createScaledBitmap(bitmap, 1024, nh, true);
-                            this.imgView.setImageBitmap(scaled);
-                        } catch (IOException e) {
+                        } catch (Exception e) {
                             e.printStackTrace();
                         }
-                        Toast.makeText(this, "이미지 선택 완료", Toast.LENGTH_SHORT).show();
-                        break;
                     default:
                         Toast.makeText(this, "취소되었습니다.", Toast.LENGTH_SHORT).show();
                         break;
                 }
-
-            }*/ else {
-                Toast.makeText(this, "알 수 없는 오류.", Toast.LENGTH_SHORT).show();
+        } else if(requestCode == checkcode[1]) {
+            switch(resultCode){
+                case RESULT_OK:
+                    absolList = data.getStringArrayListExtra("gAbsoulteFilePaths");
+                    for(int i = 0; i < absolList.size(); i++){
+                        this.storage.addImageData(absolList.get(i));
+                    }
+                    this.adapter = new ListViewAdapter(MainActivity.this,this.storage,10);
+                    this.listView.setAdapter(this.adapter);
+                    visible();
+                    Toast.makeText(this, "확인.", Toast.LENGTH_SHORT).show();
+                    break;
+                case RESULT_CANCELED:
+                    Toast.makeText(this, "취소되었습니다.", Toast.LENGTH_SHORT).show();
+                    break;
+                default:
+                    break;
             }
+        } else if(requestCode == checkcode[2]){
+            switch (resultCode){
+                case RESULT_OK:
+                    id = data.getIntExtra("id",0);
+                    cropedPath = data.getStringExtra("CropedPath");
+                    Bitmap b = BitmapFactory.decodeFile(cropedPath);
+                    View v = this.listView.getChildAt(id);
+                    ImageView imageView = v.findViewById(R.id.croped_image);
+                    imageView.setImageBitmap(b);
+                    this.storage.setCropedImage(id,b);
+                    break;
+                case RESULT_CANCELED:
+                    Toast.makeText(this,"Crop 취소",Toast.LENGTH_SHORT).show();
+                    break;
+                default:
+                    break;
+            }
+        } else {
+            Toast.makeText(this, "알 수 없는 오류.", Toast.LENGTH_SHORT).show();
+        }
     }
     @Override
     public void onBackPressed() {
         backPressCloseHandler.onBackPressed();
     }
 
-    /*
-    public String getImageApath(Uri Rpath){
-        String[] filePathColumn = { MediaStore.Images.Media.DATA };
-        Cursor cursor = getContentResolver().query(Rpath, filePathColumn, null, null, null);
-        cursor.moveToFirst();
-        Log.d("filePathColumn","filePathColumn :"+ filePathColumn);
-        int columnIndex = cursor.getColumnIndex(filePathColumn[0]);
-
-        String picturePath = cursor.getString(columnIndex);
-        Log.d("CursorCheck",
-                "getPosition :"+cursor.getPosition()+
-                        " getColumnCount :"+cursor.getColumnCount()+
-                        " getExtras :"+cursor.getExtras()+
-                        " getNtoificationUri :"+cursor.getNotificationUri()+
-                        " getColumnNames :"+cursor.getColumnName(0)
-                        //+
-               // " getColumnNames :"+cursor.getColumnName(2)+
-               // " getColumnNames :"+cursor.getColumnName(3)+
-               // " getColumnNames :"+cursor.getColumnName(4)+
-               // " getColumnNames :"+cursor.getColumnName(5)
-
-        );
-        return picturePath;
+    private void visible(){
+        textviewList[1].setVisibility(View.GONE);
+        textviewList[2].setVisibility(View.GONE);
+        buttonList[0].setVisibility(View.GONE);
+        buttonList[1].setVisibility(View.GONE);
+        buttonList[2].setVisibility(View.VISIBLE);
     }
-    */
+
+    private void writeObjectFile() throws Exception{
+        FileOutputStream fos = this.openFileOutput("adfwe!@#as",Context.MODE_PRIVATE);
+        ObjectOutputStream os = new ObjectOutputStream(fos);
+        os.writeObject(this.storage);
+        os.close();
+        fos.close();
+    }
+
+    private ListView.OnItemClickListener onItemClickListener = (adapterView, view, i, l) -> {
+        Intent intent = new Intent(MainActivity.this, CropActivity.class);
+        String s = this.storage.getAbsolutePath(i);
+        int first = this.listView.getFirstVisiblePosition();
+        int want = first-1;
+        intent.putExtra("ListViewid",want);
+        intent.putExtra("StoragePath",s);
+        startActivityForResult(intent,this.checkcode[2]);
+    };
+
+    private LinearLayout getAbsolPosition(int pos, AdapterView<?> adapterView){
+        int firstPos = adapterView.getFirstVisiblePosition();
+        int wantedPos = pos - firstPos;
+        if (wantedPos < 0 || wantedPos >= adapterView.getChildCount()) {
+            return null;
+        }
+        return (LinearLayout)adapterView.getChildAt(wantedPos);
+    }
 }
